Use the charmed-openstack-exporter snap
Upstream can take some time until releases a new snap and this is
blocking the workflow.

This change from using the golang-openstack-exporter (upstream) to
use the charmed-openstack-exporter that is packaging the upstream
code and release in the snapcraft store.

With this change the charm automatically uses the snap from the
snapstore but have preference to use the snap as resource if passed.

Functional tests will run with the snap store version, but still can
be run with a local snap if the file is provided.