Add extra-bindings for all OpenStack endpoint interface types
Add extra-bindings to the charm metadata to allow users to request legs
in all OpenStack networks spaces. This may be necessary in case the
OpenStack API endpoints to be queried are not available on the same
network space as the one used by the credentials endpoint.

While the charm doesn't explicitly use the bindingsi - meaning that the
same result could be achieved with machine space contraints - having
these specified as application bindings is more explicit.

This is effectively the same solution used in the previous exporter
charm, see
https://git.launchpad.net/charm-prometheus-openstack-exporter/commit/?id=99692c082eadbf5195bb683e92a73428603efbc6

Closes #90