Enter blocked status if resource and snap_channel
When a resource is used, the `snap_channel` config option is ignored.
In this case, the snap is installed directly from the resource file,
so a "channel" doesn't make sense.

If `snap_channel` is set and a resource is being used,
it may indicate a confusion situation,
such as the user changing the `snap_channel`
and wondering why the snap isn't being updated.
So the charm can alert the user to this situation
via a message and blocked status.

Fixes: #113