Decrease update status hook interval for func tests
This may be causing the snap down test to fail,
because the default status hook interval is 5 minutes.
Previously, we had configured the interval to be 30s,
and the status update timeout in the tests is only 2 minutes.

There is the risk that this much shorter interval will increase
the load on the VM too much because of all the extra processing.