Run func tests locally on runner (#129)
Remove the dependency on the external juju controller, switching to running them locally on a larger github runner instance. In the process, update to the latest template from our automation.

Note that I removed cinder and ceph from the func test bundle. These services are not used in the func tests, and they take up a lot of compute and storage resources because they need to run in VMs. I think removing these, plus the new larger github runners available, has enabled the tests to pass without needing to deploy on an external juju controller.