Attempt to fix functional tests
Functional tests are failing on main,
with errors from the mysql cluster.
It's potentially due to the model being too large to deploy on the runner and be stable,
or it could be related to flakiness in the mysql-innodb-cluster charm.
Either way, let's try to switch to the new mysql charm
that can be deployed in non HA mode by only deploying a single unit.
This should reduce resource use, and also perhaps be more stable.